Module Html_f

Typesafe constructors for HTML documents (Functorial interface)

See <<a_manual chapter="functors"|the manual of the functorial interface>>.

module Make : ( Xml : Xml_sigs.T with type W.ft ('a, 'b) = 'a => 'b ) => ( Svg : Svg_sigs.T with module Xml := Xml ) => Html_sigs.Make(Xml)(Svg).T with type elt (+'a) = Xml.elt and type attrib (+'a) = Xml.attrib ;

Create a new implementation of HTML, using the given underlying Xml and Svg implementation. Will output a module of type Html_sigs.T with the various type equalities.

module Wrapped_functions : ( Xml : Xml_sigs.T with type W.ft ('a, 'b) = 'a => 'b ) => Html_sigs.Wrapped_functions with module Xml = Xml ;

The standard set of wrapped functions, when W.ft is the regular function.

module Make_with_wrapped_functions : ( Xml : Xml_sigs.T ) => ( C : Html_sigs.Wrapped_functions with module Xml = Xml ) => ( Svg : Svg_sigs.T with module Xml := Xml ) => Html_sigs.Make(Xml)(Svg).T with type elt (+'a) = Xml.elt and type attrib (+'a) = Xml.attrib ;

Similar to Make but with a custom set of wrapped functions.