Module type Make.T

See Html_sigs.T .

type elt(+'a) ;

HTML elements.

Element constructors are in section elements . Most elements constructors are either nullary, unary or star, depending on the number of children they accept. Children are usually given as a list of elements. txt is used for text.

div [a [txt "Foo"]] is equivalent to <div><a>foo</a></div>

The type variable 'a is used to track the element's type. This allows the OCaml typechecker to check HTML validity.

For example, div [] is of type [> `Div] elt. The span function only accepts children of type Html_types.span_content . Since `Div is not part of it. span [div []] will not typecheck.

Note that the concrete implementation of this type can vary. See Xml for details.

type doc = elt( Html_types.html ) ;

A complete HTML document.

type attrib(+'a) ;

HTML attributes

Attribute constructors are in section attributes and their name starts with a_. Attributes are given to elements with the ~a optional argument.

a ~a:[a_href "ocsigen.org"] [txt "link!"] is equivalent to <a href="ocsigen.org">link!</a>

Similarly to elt, attributes use the OCaml type system to enforce HTML validity.

For example a_href returns a value of type [> `Href] attrib. The div function only accepts attributes of type Html_types.div_attrib . Since `Href is not part of it, div ~a:[a_href "ocsigen.org"] [] will not typecheck.

In some cases, attributes have to be disambiguated. The max attribute has two version, a_max and a_input_max , depending on the element. Such disambiguated attribute will contain the name of the associated element.

module Xml : Xml_sigs.T with type W.t ('a) = Xml.W.t ('a) with type W.tlist ('a) = Xml.W.tlist ('a) with type W.ft ('a, 'b) = Xml.W.ft ('a, 'b) with type uri = Xml.uri with type event_handler = Xml.event_handler with type mouse_event_handler = Xml.mouse_event_handler with type keyboard_event_handler = Xml.keyboard_event_handler with type touch_event_handler = Xml.touch_event_handler with type attrib = Xml.attrib with type elt = Xml.elt ;

Underlying XML data-structure

type wrap('a) = Xml.W.t ('a) ;

wrap is a container for elements and values.

In most cases, 'a wrap = 'a. For R modules (in eliom or js_of_ocaml), It will be React.S.t.

type list_wrap('a) = Xml.W.tlist ('a) ;

list_wrap is a container for list of elements.

In most cases, 'a list_wrap = 'a list. For R modules (in eliom or js_of_ocaml), It will be ReactiveData.RList.t.

type nullary('a, 'b) = ?a:list( attrib( 'a)) => unit => elt( 'b) ;

A nullary element is an element that doesn't have any children.

type unary('a, 'b, 'c) = ?a:list( attrib( 'a)) => wrap( elt( 'b)) => elt( 'c) ;

A unary element is an element that have exactly one children.

type star('a, 'b, 'c) = ?a:list( attrib( 'a)) => list_wrap (elt( 'b)) => elt( 'c) ;

A star element is an element that has any number of children, including zero.

module Info : Xml_sigs.Info ;

Various information about HTML, such as the doctype, ...

Uri

type uri = Xml.uri ;

let string_of_uri: Xml.W.ft (uri, string) ;

let uri_of_string: Xml.W.ft (string, uri) ;

Attributes

let a_class: wrap( Html_types.nmtokens ) => attrib( [> `Class ]);

This attribute assigns a class name or set of class names to an element. Any number of elements may be assigned the same class name or names.

let a_user_data: Html_types.nmtoken => wrap( Html_types.text ) => attrib( [> `User_data ]);

May be used to specify custom attributes. The example given by the W3C is as follows :

      <ol>
  <li data-length="2m11s">Beyond The Sea</li>
</ol>
     

It should be used for preprocessing ends only.

let a_id: wrap( Html_types.text ) => attrib( [> `Id ]);

This attribute assigns a name to an element. This name must be unique in a document. The text should be without any space.

let a_title: wrap( Html_types.text ) => attrib( [> `Title ]);

This attribute offers advisory information about the element for which it is set.

Values of the title attribute may be rendered by user agents in a variety of ways. For instance, visual browsers frequently display the title as a tool tip (a short message that appears when the pointing device pauses over an object). Audio user agents may speak the title information in a similar context.

The title attribute has an additional role when used with the link element to designate an external style sheet. Please consult the section on links and style sheets for details.

I18N

let a_xml_lang: wrap( Html_types.languagecode ) => attrib( [> `XML_lang ]);

let a_lang: wrap( Html_types.languagecode ) => attrib( [> `Lang ]);

Events

Javascript events

let a_onabort: Xml.event_handler => attrib( [> `OnAbort ]);

let a_onafterprint: Xml.event_handler => attrib( [> `OnAfterPrint ]);

let a_onbeforeprint: Xml.event_handler => attrib( [> `OnBeforePrint ]);

let a_onbeforeunload: Xml.event_handler => attrib( [> `OnBeforeUnload ]);

let a_onblur: Xml.event_handler => attrib( [> `OnBlur ]);

let a_oncanplay: Xml.event_handler => attrib( [> `OnCanPlay ]);

let a_oncanplaythrough: Xml.event_handler => attrib( [> `OnCanPlayThrough ]);

let a_onchange: Xml.event_handler => attrib( [> `OnChange ]);

let a_ondurationchange: Xml.event_handler => attrib( [> `OnDurationChange ]);

let a_onemptied: Xml.event_handler => attrib( [> `OnEmptied ]);

let a_onended: Xml.event_handler => attrib( [> `OnEnded ]);

let a_onerror: Xml.event_handler => attrib( [> `OnError ]);

let a_onfocus: Xml.event_handler => attrib( [> `OnFocus ]);

let a_onformchange: Xml.event_handler => attrib( [> `OnFormChange ]);

let a_onforminput: Xml.event_handler => attrib( [> `OnFormInput ]);

let a_onhashchange: Xml.event_handler => attrib( [> `OnHashChange ]);

let a_oninput: Xml.event_handler => attrib( [> `OnInput ]);

let a_oninvalid: Xml.event_handler => attrib( [> `OnInvalid ]);

let a_onmousewheel: Xml.event_handler => attrib( [> `OnMouseWheel ]);

let a_onoffline: Xml.event_handler => attrib( [> `OnOffLine ]);

let a_ononline: Xml.event_handler => attrib( [> `OnOnLine ]);

let a_onpause: Xml.event_handler => attrib( [> `OnPause ]);

let a_onplay: Xml.event_handler => attrib( [> `OnPlay ]);

let a_onplaying: Xml.event_handler => attrib( [> `OnPlaying ]);

let a_onpagehide: Xml.event_handler => attrib( [> `OnPageHide ]);

let a_onpageshow: Xml.event_handler => attrib( [> `OnPageShow ]);

let a_onpopstate: Xml.event_handler => attrib( [> `OnPopState ]);

let a_onprogress: Xml.event_handler => attrib( [> `OnProgress ]);

let a_onratechange: Xml.event_handler => attrib( [> `OnRateChange ]);

let a_onreadystatechange: Xml.event_handler => attrib( [> `OnReadyStateChange ]);

let a_onredo: Xml.event_handler => attrib( [> `OnRedo ]);

let a_onresize: Xml.event_handler => attrib( [> `OnResize ]);

let a_onscroll: Xml.event_handler => attrib( [> `OnScroll ]);

let a_onseeked: Xml.event_handler => attrib( [> `OnSeeked ]);

let a_onseeking: Xml.event_handler => attrib( [> `OnSeeking ]);

let a_onselect: Xml.event_handler => attrib( [> `OnSelect ]);

let a_onshow: Xml.event_handler => attrib( [> `OnShow ]);

let a_onstalled: Xml.event_handler => attrib( [> `OnStalled ]);

let a_onstorage: Xml.event_handler => attrib( [> `OnStorage ]);

let a_onsubmit: Xml.event_handler => attrib( [> `OnSubmit ]);

let a_onsuspend: Xml.event_handler => attrib( [> `OnSuspend ]);

let a_ontimeupdate: Xml.event_handler => attrib( [> `OnTimeUpdate ]);

let a_onundo: Xml.event_handler => attrib( [> `OnUndo ]);

let a_onunload: Xml.event_handler => attrib( [> `OnUnload ]);

let a_onvolumechange: Xml.event_handler => attrib( [> `OnVolumeChange ]);

let a_onwaiting: Xml.event_handler => attrib( [> `OnWaiting ]);

let a_onload: Xml.event_handler => attrib( [> `OnLoad ]);

let a_onloadeddata: Xml.event_handler => attrib( [> `OnLoadedData ]);

let a_onloadedmetadata: Xml.event_handler => attrib( [> `OnLoadedMetaData ]);

let a_onloadstart: Xml.event_handler => attrib( [> `OnLoadStart ]);

let a_onmessage: Xml.event_handler => attrib( [> `OnMessage ]);

Mouse events

let a_onclick: Xml.mouse_event_handler => attrib( [> `OnClick ]);

let a_oncontextmenu: Xml.mouse_event_handler => attrib( [> `OnContextMenu ]);

let a_ondblclick: Xml.mouse_event_handler => attrib( [> `OnDblClick ]);

let a_ondrag: Xml.mouse_event_handler => attrib( [> `OnDrag ]);

let a_ondragend: Xml.mouse_event_handler => attrib( [> `OnDragEnd ]);

let a_ondragenter: Xml.mouse_event_handler => attrib( [> `OnDragEnter ]);

let a_ondragleave: Xml.mouse_event_handler => attrib( [> `OnDragLeave ]);

let a_ondragover: Xml.mouse_event_handler => attrib( [> `OnDragOver ]);

let a_ondragstart: Xml.mouse_event_handler => attrib( [> `OnDragStart ]);

let a_ondrop: Xml.mouse_event_handler => attrib( [> `OnDrop ]);

let a_onmousedown: Xml.mouse_event_handler => attrib( [> `OnMouseDown ]);

let a_onmouseup: Xml.mouse_event_handler => attrib( [> `OnMouseUp ]);

let a_onmouseover: Xml.mouse_event_handler => attrib( [> `OnMouseOver ]);

let a_onmousemove: Xml.mouse_event_handler => attrib( [> `OnMouseMove ]);

let a_onmouseout: Xml.mouse_event_handler => attrib( [> `OnMouseOut ]);

let a_ontouchstart: Xml.touch_event_handler => attrib( [> `OnTouchStart ]);

Touch events

let a_ontouchend: Xml.touch_event_handler => attrib( [> `OnTouchEnd ]);

let a_ontouchmove: Xml.touch_event_handler => attrib( [> `OnTouchMove ]);

let a_ontouchcancel: Xml.touch_event_handler => attrib( [> `OnTouchCancel ]);

Keyboard events

let a_onkeypress: Xml.keyboard_event_handler => attrib( [> `OnKeyPress ]);

let a_onkeydown: Xml.keyboard_event_handler => attrib( [> `OnKeyDown ]);

let a_onkeyup: Xml.keyboard_event_handler => attrib( [> `OnKeyUp ]);

Other attributes

let a_autocomplete: wrap( bool) => attrib( [> `Autocomplete ]);

let a_async: unit => attrib( [> `Async ]);

let a_autofocus: unit => attrib( [> `Autofocus ]);

let a_autoplay: unit => attrib( [> `Autoplay ]);

let a_muted: unit => attrib( [> `Muted ]);

let a_crossorigin: wrap( [< `Anonymous | `Use_credentials ]) => attrib( [> `Crossorigin ]);

let a_mediagroup: wrap( string) => attrib( [> `Mediagroup ]);

let a_challenge: wrap( Html_types.text ) => attrib( [> `Challenge ]);

let a_contenteditable: wrap( bool) => attrib( [> `Contenteditable ]);

let a_contextmenu: wrap( Html_types.idref ) => attrib( [> `Contextmenu ]);

let a_controls: unit => attrib( [> `Controls ]);

let a_dir: wrap( [< `Rtl | `Ltr ]) => attrib( [> `Dir ]);

let a_draggable: wrap( bool) => attrib( [> `Draggable ]);

let a_form: wrap( Html_types.idref ) => attrib( [> `Form ]);

let a_formaction: wrap( Xml.uri ) => attrib( [> `Formaction ]);

let a_formenctype: wrap( Html_types.contenttype ) => attrib( [> `Formenctype ]);

let a_formnovalidate: unit => attrib( [> `Formnovalidate ]);

let a_formtarget: wrap( Html_types.text ) => attrib( [> `Formtarget ]);

let a_hidden: unit => attrib( [> `Hidden ]);

let a_high: wrap( Html_types.float_number ) => attrib( [> `High ]);

let a_icon: wrap( Xml.uri ) => attrib( [> `Icon ]);

let a_ismap: unit => attrib( [> `Ismap ]);

let a_keytype: wrap( Html_types.text ) => attrib( [> `Keytype ]);

let a_list: wrap( Html_types.idref ) => attrib( [> `List ]);

let a_loop: unit => attrib( [> `Loop ]);

let a_low: wrap( Html_types.float_number ) => attrib( [> `High ]);

let a_max: wrap( Html_types.float_number ) => attrib( [> `Max ]);

let a_input_max: wrap( Html_types.number_or_datetime ) => attrib( [> `Input_Max ]);

let a_min: wrap( Html_types.float_number ) => attrib( [> `Min ]);

let a_input_min: wrap( Html_types.number_or_datetime ) => attrib( [> `Input_Min ]);

let a_inputmode: wrap( [< `Verbatim | `Latin | `Latin_name | `Latin_prose | `Full_width_latin | `Kana | `Katakana | `Numeric | `Tel | `Email | `Url ] ) => attrib( [> `Inputmode ]);

see https://developer.mozilla.org/en-US/docs/Web/HTML/Element/input#Attributes
Input HTML documentation.

let a_novalidate: unit => attrib( [> `Novalidate ]);

let a_open: unit => attrib( [> `Open ]);

let a_optimum: wrap( Html_types.float_number ) => attrib( [> `Optimum ]);

let a_pattern: wrap( Html_types.text ) => attrib( [> `Pattern ]);

let a_placeholder: wrap( Html_types.text ) => attrib( [> `Placeholder ]);

let a_poster: wrap( Xml.uri ) => attrib( [> `Poster ]);

let a_preload: wrap( [< `None | `Metadata | `Audio ] ) => attrib( [> `Preload ]);

let a_pubdate: unit => attrib( [> `Pubdate ]);

let a_radiogroup: wrap( Html_types.text ) => attrib( [> `Radiogroup ]);

let a_required: unit => attrib( [> `Required ]);

let a_reversed: unit => attrib( [> `Reversed ]);

let a_sandbox: wrap( list( [< Html_types.sandbox_token ] )) => attrib( [> `Sandbox ]);

let a_spellcheck: wrap( bool) => attrib( [> `Spellcheck ]);

let a_scoped: unit => attrib( [> `Scoped ]);

let a_seamless: unit => attrib( [> `Seamless ]);

let a_sizes: wrap( option(list( ( Html_types.number , Html_types.number ) ))) => attrib( [> `Sizes ]);

let a_span: wrap( Html_types.number ) => attrib( [> `Span ]);

let a_srclang: wrap( Html_types.nmtoken ) => attrib( [> `XML_lang ]);

deprecated
Use a_xml_lang instead.

type image_candidate = [

| `Url( uri)

| `Url_width( (uri , Html_types.number ) )

| `Url_pixel( (uri , Html_types.float_number ) )

];

let a_srcset: wrap( list( image_candidate )) => attrib( [> `Srcset ]);

let a_img_sizes: wrap( list( Html_types.text )) => attrib( [> `Img_sizes ]);

let a_start: wrap( Html_types.number ) => attrib( [> `Start ]);

let a_step: wrap( option( Html_types.float_number )) => attrib( [> `Step ]);

let a_wrap: wrap( [< `Soft | `Hard ]) => attrib( [> `Wrap ]);

let a_version: wrap( Html_types.cdata ) => attrib( [> `Version ]);

let a_xmlns: wrap( [< `W3_org_1999_xhtml ]) => attrib( [> `XMLns ]);

let a_manifest: wrap( Xml.uri ) => attrib( [> `Manifest ]);

let a_cite: wrap( Xml.uri ) => attrib( [> `Cite ]);

let a_xml_space: wrap( [< `Default | `Preserve ]) => attrib( [> `XML_space ]);

let a_accesskey: wrap( Html_types.character ) => attrib( [> `Accesskey ]);

This attribute assigns an access key to an element. An access key is a single character from the document character set. NB: authors should consider the input method of the expected reader when specifying an accesskey.

let a_charset: wrap( Html_types.charset ) => attrib( [> `Charset ]);

This attribute specifies the character encoding of the resource designated by the link. Please consult the section on character encodings for more details.

let a_accept_charset: wrap( Html_types.charsets ) => attrib( [> `Accept_charset ]);

let a_accept: wrap( Html_types.contenttypes ) => attrib( [> `Accept ]);

let a_href: wrap( Xml.uri ) => attrib( [> `Href ]);

This attribute specifies the location of a Web resource, thus defining a link between the current element (the source anchor) and the destination anchor defined by this attribute.

let a_hreflang: wrap( Html_types.languagecode ) => attrib( [> `Hreflang ]);

This attribute specifies the base language of the resource designated by href and may only be used when href is specified.

let a_download: wrap( option(string)) => attrib( [> `Download ]);

let a_rel: wrap( Html_types.linktypes ) => attrib( [> `Rel ]);

This attribute describes the relationship from the current document to the anchor specified by the href attribute. The value of this attribute is a space-separated list of link types.

This attribute is used to describe a reverse link from the anchor specified by the href attribute to the current document. The value of this attribute is a space-separated list of link types.

let a_tabindex: wrap( Html_types.number ) => attrib( [> `Tabindex ]);

This attribute specifies the position of the current element in the tabbing order for the current document. This value must be a number between 0 and 32767. User agents should ignore leading zeros.

let a_mime_type: wrap( Html_types.contenttype ) => attrib( [> `Mime_type ]);

This attribute gives an advisory hint as to the content type of the content available at the link target address. It allows user agents to opt to use a fallback mechanism rather than fetch the content if they are advised that they will get content in a content type they do not support.Authors who use this attribute take responsibility to manage the risk that it may become inconsistent with the content available at the link target address.

let a_datetime: wrap( Html_types.cdata ) => attrib( [> `Datetime ]);

let a_action: wrap( Xml.uri ) => attrib( [> `Action ]);

This attribute specifies a form processing agent. User agent behavior for a value other than an HTTP URI is undefined.

let a_checked: unit => attrib( [> `Checked ]);

When the type attribute has the value "radio" or "checkbox", this boolean attribute specifies that the button is on. User agents must ignore this attribute for other control types.

let a_cols: wrap( Html_types.number ) => attrib( [> `Cols ]);

This attribute specifies the visible width in average character widths. Users should be able to enter longer lines than this, so user agents should provide some means to scroll through the contents of the control when the contents extend beyond the visible area. User agents may wrap visible text lines to keep long lines visible without the need for scrolling.

let a_enctype: wrap( Html_types.contenttype ) => attrib( [> `Enctype ]);

let a_label_for: wrap( Html_types.idref ) => attrib( [> `Label_for ]);

let a_for: wrap( Html_types.idref ) => attrib( [> `Label_for ]);

deprecated
Use a_label_for

let a_output_for: wrap( Html_types.idrefs ) => attrib( [> `Output_for ]);

let a_for_list: wrap( Html_types.idrefs ) => attrib( [> `Output_for ]);

deprecated
Use a_output_for

let a_maxlength: wrap( Html_types.number ) => attrib( [> `Maxlength ]);

let a_minlength: wrap( Html_types.number ) => attrib( [> `Minlength ]);

let a_method: wrap( [< `Get | `Post ]) => attrib( [> `Method ]);

let a_formmethod: wrap( [< `Get | `Post ]) => attrib( [> `Method ]);

deprecated
Use a_method

let a_multiple: unit => attrib( [> `Multiple ]);

let a_name: wrap( Html_types.text ) => attrib( [> `Name ]);

This attribute assigns the control name.

let a_rows: wrap( Html_types.number ) => attrib( [> `Rows ]);

This attribute specifies the number of visible text lines. Users should be able to enter more lines than this, so user agents should provide some means to scroll through the contents of the control when the contents extend beyond the visible area.

let a_selected: unit => attrib( [> `Selected ]);

When set, this boolean attribute specifies that this option is pre-selected.

let a_size: wrap( Html_types.number ) => attrib( [> `Size ]);

let a_src: wrap( Xml.uri ) => attrib( [> `Src ]);

let a_input_type: wrap( [< `Url | `Tel | `Text | `Time | `Search | `Password | `Checkbox | `Range | `Radio | `Submit | `Reset | `Number | `Hidden | `Month | `Week | `File | `Email | `Image | `Datetime_local | `Datetime | `Date | `Color | `Button ] ) => attrib( [> `Input_Type ]);

let a_text_value: wrap( Html_types.text ) => attrib( [> `Text_Value ]);

This attribute specifies the initial value of the control. If this attribute is not set, the initial value is set to the contents of the option element.

let a_int_value: wrap( Html_types.number ) => attrib( [> `Int_Value ]);

let a_value: wrap( Html_types.cdata ) => attrib( [> `Value ]);

let a_float_value: wrap( Html_types.float_number ) => attrib( [> `Float_Value ]);

let a_disabled: unit => attrib( [> `Disabled ]);

let a_readonly: unit => attrib( [> `ReadOnly ]);

let a_button_type: wrap( [< `Button | `Submit | `Reset ] ) => attrib( [> `Button_Type ]);

let a_command_type: wrap( [< `Command | `Checkbox | `Radio ] ) => attrib( [> `Command_Type ]);

let a_menu_type: wrap( [< `Context | `Toolbar ]) => attrib( [> `Menu_Type ]);

let a_label: wrap( Html_types.text ) => attrib( [> `Label ]);

let a_align: wrap( [< `Left | `Right | `Justify | `Char ] ) => attrib( [> `Align ]);

deprecated
Use CSS text-align

let a_axis: wrap( Html_types.cdata ) => attrib( [> `Axis ]);

deprecated
Not supported in HTML5

let a_colspan: wrap( Html_types.number ) => attrib( [> `Colspan ]);

let a_headers: wrap( Html_types.idrefs ) => attrib( [> `Headers ]);

let a_rowspan: wrap( Html_types.number ) => attrib( [> `Rowspan ]);

let a_scope: wrap( [< `Row | `Col | `Rowgroup | `Colgroup ] ) => attrib( [> `Scope ]);

deprecated
Not supported in HTML5

let a_summary: wrap( Html_types.text ) => attrib( [> `Summary ]);

deprecated
Move content elsewhere or to a <caption> child

let a_border: wrap( Html_types.pixels ) => attrib( [> `Border ]);

deprecated
Use CSS border and/or border-width

let a_rules: wrap( [< `None | `Groups | `Rows | `Cols | `All ] ) => attrib( [> `Rules ]);

deprecated
Use CSS border

let a_char: wrap( Html_types.character ) => attrib( [> `Char ]);

deprecated
The char attribute is not supported in HTML5

let a_alt: wrap( Html_types.text ) => attrib( [> `Alt ]);

let a_height: wrap( Html_types.number ) => attrib( [> `Height ]);

let a_width: wrap( Html_types.number ) => attrib( [> `Width ]);

type shape = [

| `Rect

| `Circle

| `Poly

| `Default

];

let a_shape: wrap( shape) => attrib( [> `Shape ]);

let a_coords: wrap( Html_types.numbers ) => attrib( [> `Coords ]);

let a_usemap: wrap( Html_types.idref ) => attrib( [> `Usemap ]);

let a_data: wrap( Xml.uri ) => attrib( [> `Data ]);

let a_codetype: wrap( Html_types.contenttype ) => attrib( [> `Codetype ]);

deprecated
Not supported in HTML5

let a_frameborder: wrap( [< `Zero | `One ]) => attrib( [> `Frameborder ]);

deprecated
Use CSS border

let a_marginheight: wrap( Html_types.pixels ) => attrib( [> `Marginheight ]);

deprecated
Use CSS

let a_marginwidth: wrap( Html_types.pixels ) => attrib( [> `Marginwidth ]);

deprecated
Use CSS

let a_scrolling: wrap( [< `Yes | `No | `Auto ]) => attrib( [> `Scrolling ]);

let a_target: wrap( Html_types.frametarget ) => attrib( [> `Target ]);

let a_content: wrap( Html_types.text ) => attrib( [> `Content ]);

let a_http_equiv: wrap( Html_types.text ) => attrib( [> `Http_equiv ]);

let a_defer: unit => attrib( [> `Defer ]);

let a_media: wrap( Html_types.mediadesc ) => attrib( [> `Media ]);

let a_style: wrap( string) => attrib( [> `Style_Attr ]);

let a_property: wrap( string) => attrib( [> `Property ]);

WAI-ARIA support

WAI-ARIA is a specification written by the W3C, defining a set of additional HTML attributes that can be applied to elements to provide additional semantics and improve accessibility wherever it is lacking.

See for example a WAI-ARIA tutorial .

let a_role: wrap( list(string)) => attrib( [> `Role ]);

see https://www.w3.org/TR/role-attribute
Role attribute specification

see https://www.w3.org/TR/wai-aria-1.1/#role_definitions
List of WAI-ARIA roles

let a_aria: string => wrap( list(string)) => attrib( [> `Aria ]);

Basic support for WAI-ARIA attributes: a_aria "foo" corresponds to an "aria-foo" attribute.

see https://www.w3.org/TR/wai-aria-1.1/#state_prop_def
List of WAI-ARIA attributes

Elements

let txt: wrap( string) => elt( [> Html_types.txt ] );

let html: ?a:list( attrib( Html_types.html_attrib )) => wrap( elt( [< Html_types.head ] )) => wrap( elt( [< Html_types.body ] )) => elt( [> Html_types.html ] );

let head: ?a:list( attrib( Html_types.head_attrib )) => wrap( elt( [< Html_types.title ] )) => list_wrap (elt( Html_types.head_content_fun )) => elt( [> Html_types.head ] );

let base: nullary ( [< Html_types.base_attrib ] , [> Html_types.base ] ) ;

let title: unary ( Html_types.title_attrib , [< Html_types.title_content_fun ] , [> Html_types.title ] ) ;

let body: star ( [< Html_types.body_attrib ] , [< Html_types.body_content_fun ] , [> Html_types.body ] ) ;

let svg: ?a:list( Svg.attrib ( [< Html_types.svg_attrib ] )) => list_wrap ( Svg.elt ( [< Html_types.svg_content ] )) => elt( [> Html_types.svg ] );

Section

let footer: star ( [< Html_types.footer_attrib ] , [< Html_types.footer_content_fun ] , [> Html_types.footer ] ) ;

let header: star ( [< Html_types.header_attrib ] , [< Html_types.header_content_fun ] , [> Html_types.header ] ) ;

let section: star ( [< Html_types.section_attrib ] , [< Html_types.section_content_fun ] , [> Html_types.section ] ) ;

let nav: star ( [< Html_types.nav_attrib ] , [< Html_types.nav_content_fun ] , [> Html_types.nav ] ) ;

let h1: star ( [< Html_types.h1_attrib ] , [< Html_types.h1_content_fun ] , [> Html_types.h1 ] ) ;

let h2: star ( [< Html_types.h2_attrib ] , [< Html_types.h2_content_fun ] , [> Html_types.h2 ] ) ;

let h3: star ( [< Html_types.h3_attrib ] , [< Html_types.h3_content_fun ] , [> Html_types.h3 ] ) ;

let h4: star ( [< Html_types.h4_attrib ] , [< Html_types.h4_content_fun ] , [> Html_types.h4 ] ) ;

let h5: star ( [< Html_types.h5_attrib ] , [< Html_types.h5_content_fun ] , [> Html_types.h5 ] ) ;

let h6: star ( [< Html_types.h6_attrib ] , [< Html_types.h6_content_fun ] , [> Html_types.h6 ] ) ;

let hgroup: star ( [< Html_types.hgroup_attrib ] , [< Html_types.hgroup_content_fun ] , [> Html_types.hgroup ] ) ;

let address: star ( [< Html_types.address_attrib ] , [< Html_types.address_content_fun ] , [> Html_types.address ] ) ;

let article: star ( [< Html_types.article_attrib ] , [< Html_types.article_content_fun ] , [> Html_types.article ] ) ;

let aside: star ( [< Html_types.aside_attrib ] , [< Html_types.aside_content_fun ] , [> Html_types.aside ] ) ;

let main: star ( [< Html_types.main_attrib ] , [< Html_types.main_content_fun ] , [> Html_types.main ] ) ;

Grouping content

let p: star ( [< Html_types.p_attrib ] , [< Html_types.p_content_fun ] , [> Html_types.p ] ) ;

let pre: star ( [< Html_types.pre_attrib ] , [< Html_types.pre_content_fun ] , [> Html_types.pre ] ) ;

let blockquote: star ( [< Html_types.blockquote_attrib ] , [< Html_types.blockquote_content_fun ] , [> Html_types.blockquote ] ) ;

let div: star ( [< Html_types.div_attrib ] , [< Html_types.div_content_fun ] , [> Html_types.div ] ) ;

let dl: star ( [< Html_types.dl_attrib ] , [< Html_types.dl_content_fun ] , [> Html_types.dl ] ) ;

let ol: star ( [< Html_types.ol_attrib ] , [< Html_types.ol_content_fun ] , [> Html_types.ol ] ) ;

let ul: star ( [< Html_types.ul_attrib ] , [< Html_types.ul_content_fun ] , [> Html_types.ul ] ) ;

let dd: star ( [< Html_types.dd_attrib ] , [< Html_types.dd_content_fun ] , [> Html_types.dd ] ) ;

let dt: star ( [< Html_types.dt_attrib ] , [< Html_types.dt_content_fun ] , [> Html_types.dt ] ) ;

let li: star ( [< Html_types.li_attrib ] , [< Html_types.li_content_fun ] , [> Html_types.li ] ) ;

let figcaption: star ( [< Html_types.figcaption_attrib ] , [< Html_types.figcaption_content_fun ] , [> Html_types.figcaption ] ) ;

let figure: ?figcaption: [ `Top ( wrap( elt( [< Html_types.figcaption ] ))) | `Bottom ( wrap( elt( [< Html_types.figcaption ] ))) ] => star ( [< Html_types.figure_attrib ] , [< Html_types.figure_content_fun ] , [> Html_types.figure ] ) ;

let hr: nullary ( [< Html_types.hr_attrib ] , [> Html_types.hr ] ) ;

Semantic

let b: star ( [< Html_types.b_attrib ] , [< Html_types.b_content_fun ] , [> Html_types.b ] ) ;

let i: star ( [< Html_types.i_attrib ] , [< Html_types.i_content_fun ] , [> Html_types.i ] ) ;

let u: star ( [< Html_types.u_attrib ] , [< Html_types.u_content_fun ] , [> Html_types.u ] ) ;

let small: star ( [< Html_types.small_attrib ] , [< Html_types.small_content_fun ] , [> Html_types.small ] ) ;

let sub: star ( [< Html_types.sub_attrib ] , [< Html_types.sub_content_fun ] , [> Html_types.sub ] ) ;

let sup: star ( [< Html_types.sup_attrib ] , [< Html_types.sup_content_fun ] , [> Html_types.sup ] ) ;

let mark: star ( [< Html_types.mark_attrib ] , [< Html_types.mark_content_fun ] , [> Html_types.mark ] ) ;

let wbr: nullary ( [< Html_types.wbr_attrib ] , [> Html_types.wbr ] ) ;

let bdo: dir: wrap( [< `Ltr | `Rtl ]) => star ( [< Html_types.bdo_attrib ] , [< Html_types.bdo_content_fun ] , [> Html_types.bdo ] ) ;

let abbr: star ( [< Html_types.abbr_attrib ] , [< Html_types.abbr_content_fun ] , [> Html_types.abbr ] ) ;

let br: nullary ( [< Html_types.br_attrib ] , [> Html_types.br ] ) ;

let cite: star ( [< Html_types.cite_attrib ] , [< Html_types.cite_content_fun ] , [> Html_types.cite ] ) ;

let code: star ( [< Html_types.code_attrib ] , [< Html_types.code_content_fun ] , [> Html_types.code ] ) ;

let dfn: star ( [< Html_types.dfn_attrib ] , [< Html_types.dfn_content_fun ] , [> Html_types.dfn ] ) ;

let em: star ( [< Html_types.em_attrib ] , [< Html_types.em_content_fun ] , [> Html_types.em ] ) ;

let kbd: star ( [< Html_types.kbd_attrib ] , [< Html_types.kbd_content_fun ] , [> Html_types.kbd ] ) ;

let q: star ( [< Html_types.q_attrib ] , [< Html_types.q_content_fun ] , [> Html_types.q ] ) ;

let samp: star ( [< Html_types.samp_attrib ] , [< Html_types.samp_content_fun ] , [> Html_types.samp ] ) ;

let span: star ( [< Html_types.span_attrib ] , [< Html_types.span_content_fun ] , [> Html_types.span ] ) ;

let strong: star ( [< Html_types.strong_attrib ] , [< Html_types.strong_content_fun ] , [> Html_types.strong ] ) ;

let time: star ( [< Html_types.time_attrib ] , [< Html_types.time_content_fun ] , [> Html_types.time ] ) ;

let var: star ( [< Html_types.var_attrib ] , [< Html_types.var_content_fun ] , [> Html_types.var ] ) ;

Hypertext

let a: star ( [< Html_types.a_attrib ] , 'a, [> Html_types.a ('a) ] ) ;

Edit

let del: star ( [< Html_types.del_attrib ] , 'a, [> Html_types.del ('a) ] ) ;

let ins: star ( [< Html_types.ins_attrib ] , 'a, [> Html_types.ins ('a) ] ) ;

Embedded

let img: src: wrap( Xml.uri ) => alt: wrap( Html_types.text ) => nullary ( [< Html_types.img_attrib ] , [> Html_types.img ] ) ;

let iframe: star ( [< Html_types.iframe_attrib ] , [< Html_types.iframe_content_fun ] , [> Html_types.iframe ] ) ;

let object_: ?params: list_wrap (elt( [< Html_types.param ] )) => star ( [< Html_types.object__attrib ] , 'a, [> `Object('a) ] ) ;

let param: nullary ( [< Html_types.param_attrib ] , [> Html_types.param ] ) ;

let embed: nullary ( [< Html_types.embed_attrib ] , [> Html_types.embed ] ) ;

let audio: ?src: wrap( Xml.uri ) => ?srcs: list_wrap (elt( [< Html_types.source ] )) => star ( [< Html_types.audio_attrib ] , 'a, [> Html_types.audio ('a) ] ) ;

let video: ?src: wrap( Xml.uri ) => ?srcs: list_wrap (elt( [< Html_types.source ] )) => star ( [< Html_types.video_attrib ] , 'a, [> Html_types.video ('a) ] ) ;

let canvas: star ( [< Html_types.canvas_attrib ] , 'a, [> Html_types.canvas ('a) ] ) ;

let source: nullary ( [< Html_types.source_attrib ] , [> Html_types.source ] ) ;

let area: alt: wrap( Html_types.text ) => nullary ( [< Html_types.common | `Alt | `Coords | `Shape | `Target | `Rel | `Media | `Hreflang | `Mime_type ] , [> Html_types.area ] ) ;

let map: star ( [< Html_types.map_attrib ] , 'a, [> Html_types.map ('a) ] ) ;

Tables Data

let caption: star ( [< Html_types.caption_attrib ] , [< Html_types.caption_content_fun ] , [> Html_types.caption ] ) ;

let table: ?caption: wrap( elt( [< Html_types.caption ] )) => ?columns: list_wrap (elt( [< Html_types.colgroup ] )) => ?thead: wrap( elt( [< Html_types.thead ] )) => ?tfoot: wrap( elt( [< Html_types.tfoot ] )) => star ( [< Html_types.table_attrib ] , [< Html_types.table_content_fun ] , [> Html_types.table ] ) ;

let tablex: ?caption: wrap( elt( [< Html_types.caption ] )) => ?columns: list_wrap (elt( [< Html_types.colgroup ] )) => ?thead: wrap( elt( [< Html_types.thead ] )) => ?tfoot: wrap( elt( [< Html_types.tfoot ] )) => star ( [< Html_types.tablex_attrib ] , [< Html_types.tablex_content_fun ] , [> Html_types.tablex ] ) ;

let colgroup: star ( [< Html_types.colgroup_attrib ] , [< Html_types.colgroup_content_fun ] , [> Html_types.colgroup ] ) ;

let col: nullary ( [< Html_types.col_attrib ] , [> Html_types.col ] ) ;

let thead: star ( [< Html_types.thead_attrib ] , [< Html_types.thead_content_fun ] , [> Html_types.thead ] ) ;

let tbody: star ( [< Html_types.tbody_attrib ] , [< Html_types.tbody_content_fun ] , [> Html_types.tbody ] ) ;

let tfoot: star ( [< Html_types.tfoot_attrib ] , [< Html_types.tfoot_content_fun ] , [> Html_types.tfoot ] ) ;

let td: star ( [< Html_types.td_attrib ] , [< Html_types.td_content_fun ] , [> Html_types.td ] ) ;

let th: star ( [< Html_types.th_attrib ] , [< Html_types.th_content_fun ] , [> Html_types.th ] ) ;

let tr: star ( [< Html_types.tr_attrib ] , [< Html_types.tr_content_fun ] , [> Html_types.tr ] ) ;

Forms

let form: star ( [< Html_types.form_attrib ] , [< Html_types.form_content_fun ] , [> Html_types.form ] ) ;

let fieldset: ?legend: wrap( elt( [< Html_types.legend ] )) => star ( [< Html_types.fieldset_attrib ] , [< Html_types.fieldset_content_fun ] , [> Html_types.fieldset ] ) ;

let legend: star ( [< Html_types.legend_attrib ] , [< Html_types.legend_content_fun ] , [> Html_types.legend ] ) ;

let label: star ( [< Html_types.label_attrib ] , [< Html_types.label_content_fun ] , [> Html_types.label ] ) ;

Label authorizes only one control inside them that should be labelled with a for attribute (although it is not necessary). Such constraints are not currently enforced by the type-system

let input: nullary ( [< Html_types.input_attrib ] , [> Html_types.input ] ) ;

let button: star ( [< Html_types.button_attrib ] , [< Html_types.button_content_fun ] , [> Html_types.button ] ) ;

let select: star ( [< Html_types.select_attrib ] , [< Html_types.select_content_fun ] , [> Html_types.select ] ) ;

let datalist: ?children: [< `Options ( list_wrap (elt( [< Html_types.selectoption ] ))) | `Phras ( list_wrap (elt( [< Html_types.phrasing ] ))) ] => nullary ( [< Html_types.datalist_attrib ] , [> Html_types.datalist ] ) ;

let optgroup: label: wrap( Html_types.text ) => star ( [< Html_types.optgroup_attrib ] , [< Html_types.optgroup_content_fun ] , [> Html_types.optgroup ] ) ;

let option: unary ( [< Html_types.option_attrib ] , [< Html_types.option_content_fun ] , [> Html_types.selectoption ] ) ;

let textarea: unary ( [< Html_types.textarea_attrib ] , [< Html_types.textarea_content_fun ] , [> Html_types.textarea ] ) ;

let keygen: nullary ( [< Html_types.keygen_attrib ] , [> Html_types.keygen ] ) ;

let progress: star ( [< Html_types.progress_attrib ] , [< Html_types.progress_content_fun ] , [> Html_types.progress ] ) ;

let meter: star ( [< Html_types.meter_attrib ] , [< Html_types.meter_content_fun ] , [> Html_types.meter ] ) ;

let output_elt: star ( [< Html_types.output_elt_attrib ] , [< Html_types.output_elt_content_fun ] , [> Html_types.output_elt ] ) ;

Data

let entity: string => elt( [> Html_types.txt ] );

entity "foo" is the HTML entity &foo;. Both numerical and named form are allowed.

see http://www.w3schools.com/html/html_entities.asp
A tutorial on HTML entities.

see https://www.w3.org/TR/html5/syntax.html#named-character-references
The list of HTML entities.

let space: unit => elt( [> Html_types.txt ] );

let cdata: string => elt( [> Html_types.txt ] );

let cdata_script: string => elt( [> Html_types.txt ] );

let cdata_style: string => elt( [> Html_types.txt ] );

Interactive

let details: wrap( elt( [< Html_types.summary ] )) => star ( [< Html_types.details_attrib ] , [< Html_types.details_content_fun ] , [> Html_types.details ] ) ;

let summary: star ( [< Html_types.summary_attrib ] , [< Html_types.summary_content_fun ] , [> Html_types.summary ] ) ;

let command: label: wrap( Html_types.text ) => nullary ( [< Html_types.command_attrib ] , [> Html_types.command ] ) ;

let menu: ?children: [< `Lis ( list_wrap (elt( [< `Li ( [< Html_types.common ] ) ] ))) | `Flows ( list_wrap (elt( [< Html_types.flow5 ] ))) ] => nullary ( [< Html_types.menu_attrib ] , [> Html_types.menu ] ) ;

Scripting

let script: unary ( [< Html_types.script_attrib ] , [< Html_types.script_content_fun ] , [> Html_types.script ] ) ;

let noscript: star ( [< Html_types.noscript_attrib ] , [< Html_types.noscript_content_fun ] , [> Html_types.noscript ] ) ;

let meta: nullary ( [< Html_types.meta_attrib ] , [> Html_types.meta ] ) ;

Style Sheets

let style: star ( [< Html_types.style_attrib ] , [< Html_types.style_content_fun ] , [> Html_types.style ] ) ;

let link: rel: wrap( Html_types.linktypes ) => href: wrap( Xml.uri ) => nullary ( [< Html_types.link_attrib ] , [> Html_types.link ] ) ;

Ruby

let rt: star ( [< Html_types.rt_attrib ] , [< Html_types.rt_content_fun ] , [> Html_types.rt ] ) ;

let rp: star ( [< Html_types.rp_attrib ] , [< Html_types.rp_content_fun ] , [> Html_types.rp ] ) ;

let ruby: star ( [< Html_types.ruby_attrib ] , [< Html_types.ruby_content_fun ] , [> Html_types.ruby ] ) ;

Deprecated

let pcdata: wrap( string) => elt( [> Html_types.pcdata ] );

deprecated
Use txt instead

Conversion with untyped representation

WARNING: These functions do not ensure HTML or SVG validity! You should always explicitly given an appropriate type to the output.

let of_seq: Stdlib.Seq.t( Xml_stream.signal ) => list_wrap (elt( 'a));

import signal converts the given XML signal into Tyxml elements. It can be used with HTML and SVG parsing libraries, such as Markup.

raises Xml_stream.Malformed_stream
if the stream is malformed.

let tot: Xml.elt => elt( 'a);

let totl: list_wrap ( Xml.elt ) => list_wrap (elt( 'a));

let toelt: elt( 'a) => Xml.elt ;

let toeltl: list_wrap (elt( 'a)) => list_wrap ( Xml.elt );

let doc_toelt: doc => Xml.elt ;

let to_xmlattribs: list( attrib( 'a)) => list( Xml.attrib );

let to_attrib: Xml.attrib => attrib( 'a);

module Unsafe : { ... };

Unsafe features.