Module T.Unsafe

Unsafe features.

Using this module can break HTML validity and may introduce security problems like code injection. Use it with care.

let data: wrap (string) => elt( 'a);

Insert raw text without any encoding

let node: string => ?a:list( attrib ('a)) => list_wrap ( elt( 'b)) => elt( 'c);

Insert an XML node that is not implemented in this module. If it is a standard HTML node which is missing, please report to the Ocsigen team.

let leaf: string => ?a:list( attrib ('a)) => unit => elt( 'b);

Insert an XML node without children that is not implemented in this module. If it is a standard HTML node which is missing, please report to the Ocsigen team.

let coerce_elt: elt( 'a) => elt( 'b);

Remove phantom type annotation on an element, to make it usable everywhere.

let string_attrib: string => wrap (string) => attrib ('a);

Insert an attribute that is not implemented in this module. If it is a standard HTML attribute which is missing, please report to the Ocsigen team.

let float_attrib: string => wrap (float) => attrib ('a);

Same, for float attribute

let int_attrib: string => wrap (int) => attrib ('a);

Same, for int attribute

let uri_attrib: string => wrap (uri) => attrib ('a);

Same, for URI attribute

let space_sep_attrib: string => wrap (list(string)) => attrib ('a);

Same, for a space separated list of values

let comma_sep_attrib: string => wrap (list(string)) => attrib ('a);

Same, for a comma separated list of values