Module Svg_f.Make

Create a new implementation of Svg, using the given underlying Xml implementation. Will output a module of type Svg_sigs.T with the various type equalities.

If your Xml implementation uses a special function wrapping, use Make_with_wrapped_functions .

Parameters

module Xml : Xml_sigs.T with type W.ft ('a, 'b) = 'a => 'b

Signature

type elt(+'a) = Xml.elt ;

SVG elements.

Element constructors are in section Elements. Most elements constructors are either nullary, unary or star, depending on the number of children they accept. Children are usually given as a list of elements. txt is used for text.

The type variable 'a is used to track the element's type. This allows the OCaml typechecker to check SVG validity.

Note that the concrete implementation of this type can vary. See Xml for details.

type doc = elt( [ `Svg ]) ;

A complete SVG document.

type attrib(+'a) = Xml.attrib ;

SVG attributes

Attribute constructors are in section Attributes and their name starts with a_. Attributes are given to elements with the ~a optional argument.

Similarly to elt, attributes use the OCaml type system to enforce Html validity.

In some cases, attributes have to be disambiguated. The max attribute has two version, a_fill and a_animation_fill , depending on the element. Such disambiguated attribute will contain the name of the associated element.

module Xml : Xml_sigs.T with type W.t ('a) = Xml.W.t ('a) with type W.tlist ('a) = Xml.W.tlist ('a) with type W.ft ('a, 'b) = Xml.W.ft ('a, 'b) with type uri = Xml.uri with type event_handler = Xml.event_handler with type mouse_event_handler = Xml.mouse_event_handler with type keyboard_event_handler = Xml.keyboard_event_handler with type touch_event_handler = Xml.touch_event_handler with type attrib = Xml.attrib with type elt = Xml.elt ;

Underlying XML data-structure

type wrap('a) = Xml.W.t ('a) ;

wrap is a container for elements and values.

In most cases, 'a wrap = 'a. For R modules (in eliom or js_of_ocaml), It will be React.S.t.

type list_wrap('a) = Xml.W.tlist ('a) ;

list_wrap is a containre for list of elements.

In most cases, 'a list_wrap = 'a list. For R modules (in eliom or js_of_ocaml), It will be ReactiveData.RList.t.

type nullary('a, 'b) = ?a:list( attrib( 'a)) => unit => elt( 'b) ;

A nullary element is an element that doesn't have any children.

type unary('a, 'b, 'c) = ?a:list( attrib( 'a)) => wrap( elt( 'b)) => elt( 'c) ;

A unary element is an element that have exactly one children.

type star('a, 'b, 'c) = ?a:list( attrib( 'a)) => list_wrap( elt( 'b)) => elt( 'c) ;

A star element is an element that has any number of children, including zero.

module Info : Xml_sigs.Info ;

Various information about SVG, such as the doctype, ...

Uri

type uri = Xml.uri ;

let string_of_uri: Xml.W.ft (uri, string) ;

let uri_of_string: Xml.W.ft (string, uri) ;

Attributes

let a_version: wrap(string) => attrib( [> `Version ]);

deprecated
Removed in SVG2

let a_baseProfile: wrap(string) => attrib( [> `BaseProfile ]);

deprecated
Removed in SVG2

let a_x: wrap( Svg_types.coord ) => attrib( [> `X ]);

let a_y: wrap( Svg_types.coord ) => attrib( [> `Y ]);

let a_width: wrap( Svg_types.Unit.length ) => attrib( [> `Width ]);

let a_height: wrap( Svg_types.Unit.length ) => attrib( [> `Height ]);

let a_preserveAspectRatio: wrap(string) => attrib( [> `PreserveAspectRatio ]);

let a_contentScriptType: wrap(string) => attrib( [> `ContentScriptType ]);

deprecated
Removed in SVG2

let a_contentStyleType: wrap(string) => attrib( [> `ContentStyleType ]);

deprecated
Removed in SVG2

let a_zoomAndPan: wrap( [< `Disable | `Magnify ]) => attrib( [> `ZoomAndSpan ]);

let a_href: wrap( Svg_types.iri ) => attrib( [> `Xlink_href ]);

let a_xlink_href: wrap( Svg_types.iri ) => attrib( [> `Xlink_href ]);

deprecated
Use a_href

let a_requiredFeatures: wrap( Svg_types.spacestrings ) => attrib( [> `RequiredFeatures ]);

deprecated
Removed in SVG2

let a_requiredExtensions: wrap( Svg_types.spacestrings ) => attrib( [> `RequiredExtension ]);

let a_systemLanguage: wrap( Svg_types.commastrings ) => attrib( [> `SystemLanguage ]);

let a_externalRessourcesRequired: wrap(bool) => attrib( [> `ExternalRessourcesRequired ]);

let a_id: wrap(string) => attrib( [> `Id ]);

let a_xml_base: wrap( Svg_types.iri ) => attrib( [> `Xml_Base ]);

deprecated
Removed in SVG2

let a_xml_lang: wrap( Svg_types.iri ) => attrib( [> `Xml_Lang ]);

let a_xml_space: wrap( [< `Default | `Preserve ]) => attrib( [> `Xml_Space ]);

deprecated
Use CSS white-space

let a_type: wrap(string) => attrib( [> `Type ]);

let a_media: wrap( Svg_types.commastrings ) => attrib( [> `Media ]);

let a_xlink_title: wrap(string) => attrib( [> `Title ]);

deprecated
Use a child title element

let a_class: wrap( Svg_types.spacestrings ) => attrib( [> `Class ]);

let a_style: wrap(string) => attrib( [> `Style ]);

let a_transform: wrap( Svg_types.transforms ) => attrib( [> `Transform ]);

let a_viewBox: wrap( Svg_types.fourfloats ) => attrib( [> `ViewBox ]);

let a_d: wrap(string) => attrib( [> `D ]);

let a_pathLength: wrap(float) => attrib( [> `PathLength ]);

let a_rx: wrap( Svg_types.Unit.length ) => attrib( [> `Rx ]);

let a_ry: wrap( Svg_types.Unit.length ) => attrib( [> `Ry ]);

let a_cx: wrap( Svg_types.Unit.length ) => attrib( [> `Cx ]);

let a_cy: wrap( Svg_types.Unit.length ) => attrib( [> `Cy ]);

let a_r: wrap( Svg_types.Unit.length ) => attrib( [> `R ]);

let a_x1: wrap( Svg_types.coord ) => attrib( [> `X1 ]);

let a_y1: wrap( Svg_types.coord ) => attrib( [> `Y1 ]);

let a_x2: wrap( Svg_types.coord ) => attrib( [> `X2 ]);

let a_y2: wrap( Svg_types.coord ) => attrib( [> `Y2 ]);

let a_points: wrap( Svg_types.coords ) => attrib( [> `Points ]);

let a_x_list: wrap( Svg_types.lengths ) => attrib( [> `X_list ]);

let a_y_list: wrap( Svg_types.lengths ) => attrib( [> `Y_list ]);

let a_dx: wrap( Svg_types.number ) => attrib( [> `Dx ]);

let a_dy: wrap( Svg_types.number ) => attrib( [> `Dy ]);

let a_dx_list: wrap( Svg_types.lengths ) => attrib( [> `Dx_list ]);

let a_dy_list: wrap( Svg_types.lengths ) => attrib( [> `Dy_list ]);

let a_lengthAdjust: wrap( [< `Spacing | `SpacingAndGlyphs ]) => attrib( [> `LengthAdjust ]);

let a_textLength: wrap( Svg_types.Unit.length ) => attrib( [> `TextLength ]);

let a_text_anchor: wrap( [< `Start | `Middle | `End | `Inherit ] ) => attrib( [> `Text_Anchor ]);

let a_text_decoration: wrap( [< `None | `Underline | `Overline | `Line_through | `Blink | `Inherit ] ) => attrib( [> `Text_Decoration ]);

let a_text_rendering: wrap( [< `Auto | `OptimizeSpeed | `OptimizeLegibility | `GeometricPrecision | `Inherit ] ) => attrib( [> `Text_Rendering ]);

let a_rotate: wrap( Svg_types.numbers ) => attrib( [> `Rotate ]);

let a_startOffset: wrap( Svg_types.Unit.length ) => attrib( [> `StartOffset ]);

let a_method: wrap( [< `Align | `Stretch ]) => attrib( [> `Method ]);

let a_spacing: wrap( [< `Auto | `Exact ]) => attrib( [> `Spacing ]);

let a_glyphRef: wrap(string) => attrib( [> `GlyphRef ]);

let a_format: wrap(string) => attrib( [> `Format ]);

let a_markerUnits: wrap( [< `StrokeWidth | `UserSpaceOnUse ]) => attrib( [> `MarkerUnits ]);

let a_refX: wrap( Svg_types.coord ) => attrib( [> `RefX ]);

let a_refY: wrap( Svg_types.coord ) => attrib( [> `RefY ]);

let a_markerWidth: wrap( Svg_types.Unit.length ) => attrib( [> `MarkerWidth ]);

let a_markerHeight: wrap( Svg_types.Unit.length ) => attrib( [> `MarkerHeight ]);

let a_orient: wrap(option( Svg_types.Unit.angle )) => attrib( [> `Orient ]);

let a_local: wrap(string) => attrib( [> `Local ]);

let a_rendering_intent: wrap( [< `Auto | `Perceptual | `Relative_colorimetric | `Saturation | `Absolute_colorimetric ] ) => attrib( [> `Rendering_Indent ]);

let a_gradientUnits: wrap( [< `UserSpaceOnUse | `ObjectBoundingBox ] ) => attrib( [ `GradientUnits ]);

let a_gradientTransform: wrap( Svg_types.transforms ) => attrib( [> `Gradient_Transform ]);

let a_spreadMethod: wrap( [< `Pad | `Reflect | `Repeat ] ) => attrib( [> `SpreadMethod ]);

let a_fx: wrap( Svg_types.coord ) => attrib( [> `Fx ]);

let a_fy: wrap( Svg_types.coord ) => attrib( [> `Fy ]);

let a_offset: wrap( [< `Number ( Svg_types.number ) | `Percentage ( Svg_types.percentage ) ] ) => attrib( [> `Offset ]);

let a_patternUnits: wrap( [< `UserSpaceOnUse | `ObjectBoundingBox ] ) => attrib( [> `PatternUnits ]);

let a_patternContentUnits: wrap( [< `UserSpaceOnUse | `ObjectBoundingBox ] ) => attrib( [> `PatternContentUnits ]);

let a_patternTransform: wrap( Svg_types.transforms ) => attrib( [> `PatternTransform ]);

let a_clipPathUnits: wrap( [< `UserSpaceOnUse | `ObjectBoundingBox ] ) => attrib( [> `ClipPathUnits ]);

let a_maskUnits: wrap( [< `UserSpaceOnUse | `ObjectBoundingBox ] ) => attrib( [> `MaskUnits ]);

let a_maskContentUnits: wrap( [< `UserSpaceOnUse | `ObjectBoundingBox ] ) => attrib( [> `MaskContentUnits ]);

let a_primitiveUnits: wrap( [< `UserSpaceOnUse | `ObjectBoundingBox ] ) => attrib( [> `PrimitiveUnits ]);

let a_filterRes: wrap( Svg_types.number_optional_number ) => attrib( [> `FilterResUnits ]);

let a_result: wrap(string) => attrib( [> `Result ]);

let a_in: wrap( [< `SourceGraphic | `SourceAlpha | `BackgroundImage | `BackgroundAlpha | `FillPaint | `StrokePaint | `Ref(string) ] ) => attrib( [> `In ]);

let a_in2: wrap( [< `SourceGraphic | `SourceAlpha | `BackgroundImage | `BackgroundAlpha | `FillPaint | `StrokePaint | `Ref(string) ] ) => attrib( [> `In2 ]);

let a_azimuth: wrap(float) => attrib( [> `Azimuth ]);

let a_elevation: wrap(float) => attrib( [> `Elevation ]);

let a_pointsAtX: wrap(float) => attrib( [> `PointsAtX ]);

let a_pointsAtY: wrap(float) => attrib( [> `PointsAtY ]);

let a_pointsAtZ: wrap(float) => attrib( [> `PointsAtZ ]);

let a_specularExponent: wrap(float) => attrib( [> `SpecularExponent ]);

let a_specularConstant: wrap(float) => attrib( [> `SpecularConstant ]);

let a_limitingConeAngle: wrap(float) => attrib( [> `LimitingConeAngle ]);

let a_mode: wrap( [< `Normal | `Multiply | `Screen | `Darken | `Lighten ] ) => attrib( [> `Mode ]);

let a_feColorMatrix_type: wrap( [< `Matrix | `Saturate | `HueRotate | `LuminanceToAlpha ] ) => attrib( [> `Typefecolor ]);

let a_values: wrap( Svg_types.numbers ) => attrib( [> `Values ]);

let a_transfer_type: wrap( [< `Identity | `Table | `Discrete | `Linear | `Gamma ] ) => attrib( [> `Type_transfert ]);

let a_tableValues: wrap( Svg_types.numbers ) => attrib( [> `TableValues ]);

let a_intercept: wrap( Svg_types.number ) => attrib( [> `Intercept ]);

let a_amplitude: wrap( Svg_types.number ) => attrib( [> `Amplitude ]);

let a_exponent: wrap( Svg_types.number ) => attrib( [> `Exponent ]);

let a_transfer_offset: wrap( Svg_types.number ) => attrib( [> `Offset_transfer ]);

let a_feComposite_operator: wrap( [< `Over | `In | `Out | `Atop | `Xor | `Arithmetic ] ) => attrib( [> `OperatorComposite ]);

let a_k1: wrap( Svg_types.number ) => attrib( [> `K1 ]);

let a_k2: wrap( Svg_types.number ) => attrib( [> `K2 ]);

let a_k3: wrap( Svg_types.number ) => attrib( [> `K3 ]);

let a_k4: wrap( Svg_types.number ) => attrib( [> `K4 ]);

let a_order: wrap( Svg_types.number_optional_number ) => attrib( [> `Order ]);

let a_kernelMatrix: wrap( Svg_types.numbers ) => attrib( [> `KernelMatrix ]);

let a_divisor: wrap( Svg_types.number ) => attrib( [> `Divisor ]);

let a_bias: wrap( Svg_types.number ) => attrib( [> `Bias ]);

let a_kernelUnitLength: wrap( Svg_types.number_optional_number ) => attrib( [> `KernelUnitLength ]);

let a_targetX: wrap(int) => attrib( [> `TargetX ]);

let a_targetY: wrap(int) => attrib( [> `TargetY ]);

let a_edgeMode: wrap( [< `Duplicate | `Wrap | `None ] ) => attrib( [> `TargetY ]);

let a_preserveAlpha: wrap(bool) => attrib( [> `TargetY ]);

let a_surfaceScale: wrap( Svg_types.number ) => attrib( [> `SurfaceScale ]);

let a_diffuseConstant: wrap( Svg_types.number ) => attrib( [> `DiffuseConstant ]);

let a_scale: wrap( Svg_types.number ) => attrib( [> `Scale ]);

let a_xChannelSelector: wrap( [< `R | `G | `B | `A ] ) => attrib( [> `XChannelSelector ]);

let a_yChannelSelector: wrap( [< `R | `G | `B | `A ] ) => attrib( [> `YChannelSelector ]);

let a_stdDeviation: wrap( Svg_types.number_optional_number ) => attrib( [> `StdDeviation ]);

let a_feMorphology_operator: wrap( [< `Erode | `Dilate ]) => attrib( [> `OperatorMorphology ]);

let a_radius: wrap( Svg_types.number_optional_number ) => attrib( [> `Radius ]);

let a_baseFrenquency: wrap( Svg_types.number_optional_number ) => attrib( [> `BaseFrequency ]);

let a_numOctaves: wrap(int) => attrib( [> `NumOctaves ]);

let a_seed: wrap( Svg_types.number ) => attrib( [> `Seed ]);

let a_stitchTiles: wrap( [< `Stitch | `NoStitch ]) => attrib( [> `StitchTiles ]);

let a_feTurbulence_type: wrap( [< `FractalNoise | `Turbulence ]) => attrib( [> `TypeStitch ]);

let a_xlink_show: wrap( [< `New | `Replace ]) => attrib( [> `Xlink_show ]);

deprecated
Removed in SVG2

let a_xlink_actuate: wrap( [< `OnRequest | `OnLoad | `Other | `None ] ) => attrib( [> `Xlink_actuate ]);

deprecated
Removed in SVG2

let a_target: wrap(string) => attrib( [> `Xlink_target ]);

let a_viewTarget: wrap(string) => attrib( [> `ViewTarget ]);

deprecated
Removed in SVG2

let a_attributeName: wrap(string) => attrib( [> `AttributeName ]);

let a_attributeType: wrap( [< `CSS | `XML | `Auto ]) => attrib( [> `AttributeType ]);

let a_begin: wrap(string) => attrib( [> `Begin ]);

let a_dur: wrap(string) => attrib( [> `Dur ]);

let a_min: wrap(string) => attrib( [> `Min ]);

let a_max: wrap(string) => attrib( [> `Max ]);

let a_restart: wrap( [< `Always | `WhenNotActive | `Never ] ) => attrib( [> `Restart ]);

let a_repeatCount: wrap(string) => attrib( [> `RepeatCount ]);

let a_repeatDur: wrap(string) => attrib( [> `RepeatDur ]);

let a_fill: wrap( Svg_types.paint ) => attrib( [> `Fill ]);

let a_animation_fill: wrap( [< `Freeze | `Remove ]) => attrib( [> `Fill_Animation ]);

let a_calcMode: wrap( [< `Discrete | `Linear | `Paced | `Spline ] ) => attrib( [> `CalcMode ]);

let a_animation_values: wrap( Svg_types.strings ) => attrib( [> `Valuesanim ]);

let a_keyTimes: wrap( Svg_types.strings ) => attrib( [> `KeyTimes ]);

let a_keySplines: wrap( Svg_types.strings ) => attrib( [> `KeySplines ]);

let a_from: wrap(string) => attrib( [> `From ]);

let a_to: wrap(string) => attrib( [> `To ]);

let a_by: wrap(string) => attrib( [> `By ]);

let a_additive: wrap( [< `Replace | `Sum ]) => attrib( [> `Additive ]);

let a_accumulate: wrap( [< `None | `Sum ]) => attrib( [> `Accumulate ]);

let a_keyPoints: wrap( Svg_types.numbers_semicolon ) => attrib( [> `KeyPoints ]);

let a_path: wrap(string) => attrib( [> `Path ]);

let a_animateTransform_type: wrap( [ `Translate | `Scale | `Rotate | `SkewX | `SkewY ] ) => attrib( [ `Typeanimatetransform ]);

let a_horiz_origin_x: wrap( Svg_types.number ) => attrib( [> `HorizOriginX ]);

let a_horiz_origin_y: wrap( Svg_types.number ) => attrib( [> `HorizOriginY ]);

let a_horiz_adv_x: wrap( Svg_types.number ) => attrib( [> `HorizAdvX ]);

let a_vert_origin_x: wrap( Svg_types.number ) => attrib( [> `VertOriginX ]);

let a_vert_origin_y: wrap( Svg_types.number ) => attrib( [> `VertOriginY ]);

let a_vert_adv_y: wrap( Svg_types.number ) => attrib( [> `VertAdvY ]);

let a_unicode: wrap(string) => attrib( [> `Unicode ]);

let a_glyph_name: wrap(string) => attrib( [> `glyphname ]);

let a_orientation: wrap( [< `H | `V ]) => attrib( [> `Orientation ]);

let a_arabic_form: wrap( [< `Initial | `Medial | `Terminal | `Isolated ] ) => attrib( [> `Arabicform ]);

let a_lang: wrap(string) => attrib( [> `Lang ]);

let a_u1: wrap(string) => attrib( [> `U1 ]);

let a_u2: wrap(string) => attrib( [> `U2 ]);

let a_g1: wrap(string) => attrib( [> `G1 ]);

let a_g2: wrap(string) => attrib( [> `G2 ]);

let a_k: wrap(string) => attrib( [> `K ]);

let a_font_family: wrap(string) => attrib( [> `Font_Family ]);

let a_font_style: wrap(string) => attrib( [> `Font_Style ]);

let a_font_variant: wrap(string) => attrib( [> `Font_Variant ]);

let a_font_weight: wrap(string) => attrib( [> `Font_Weight ]);

let a_font_stretch: wrap(string) => attrib( [> `Font_Stretch ]);

let a_font_size: wrap(string) => attrib( [> `Font_Size ]);

let a_unicode_range: wrap(string) => attrib( [> `UnicodeRange ]);

let a_units_per_em: wrap(string) => attrib( [> `UnitsPerEm ]);

let a_stemv: wrap( Svg_types.number ) => attrib( [> `Stemv ]);

let a_stemh: wrap( Svg_types.number ) => attrib( [> `Stemh ]);

let a_slope: wrap( Svg_types.number ) => attrib( [> `Slope ]);

let a_cap_height: wrap( Svg_types.number ) => attrib( [> `CapHeight ]);

let a_x_height: wrap( Svg_types.number ) => attrib( [> `XHeight ]);

let a_accent_height: wrap( Svg_types.number ) => attrib( [> `AccentHeight ]);

let a_ascent: wrap( Svg_types.number ) => attrib( [> `Ascent ]);

let a_widths: wrap(string) => attrib( [> `Widths ]);

let a_bbox: wrap(string) => attrib( [> `Bbox ]);

let a_ideographic: wrap( Svg_types.number ) => attrib( [> `Ideographic ]);

let a_alphabetic: wrap( Svg_types.number ) => attrib( [> `Alphabetic ]);

let a_mathematical: wrap( Svg_types.number ) => attrib( [> `Mathematical ]);

let a_hanging: wrap( Svg_types.number ) => attrib( [> `Hanging ]);

let a_videographic: wrap( Svg_types.number ) => attrib( [> `VIdeographic ]);

let a_v_alphabetic: wrap( Svg_types.number ) => attrib( [> `VAlphabetic ]);

let a_v_mathematical: wrap( Svg_types.number ) => attrib( [> `VMathematical ]);

let a_v_hanging: wrap( Svg_types.number ) => attrib( [> `VHanging ]);

let a_underline_position: wrap( Svg_types.number ) => attrib( [> `UnderlinePosition ]);

let a_underline_thickness: wrap( Svg_types.number ) => attrib( [> `UnderlineThickness ]);

let a_strikethrough_position: wrap( Svg_types.number ) => attrib( [> `StrikethroughPosition ]);

let a_strikethrough_thickness: wrap( Svg_types.number ) => attrib( [> `StrikethroughThickness ]);

let a_overline_position: wrap( Svg_types.number ) => attrib( [> `OverlinePosition ]);

let a_overline_thickness: wrap( Svg_types.number ) => attrib( [> `OverlineThickness ]);

let a_string: wrap(string) => attrib( [> `String ]);

let a_name: wrap(string) => attrib( [> `Name ]);

let a_alignment_baseline: wrap( [< `Auto | `Baseline | `Before_edge | `Text_before_edge | `Middle | `Central | `After_edge | `Text_after_edge | `Ideographic | `Alphabetic | `Hanging | `Mathematical | `Inherit ] ) => attrib( [> `Alignment_Baseline ]);

let a_dominant_baseline: wrap( [< `Auto | `Use_script | `No_change | `Reset_size | `Ideographic | `Alphabetic | `Hanging | `Mathematical | `Central | `Middle | `Text_after_edge | `Text_before_edge | `Inherit ] ) => attrib( [> `Dominant_Baseline ]);

let a_stop_color: wrap( Svg_types.color ) => attrib( [> `Stop_Color ]);

let a_stop_opacity: wrap( Svg_types.number ) => attrib( [> `Stop_Opacity ]);

let a_stroke: wrap( Svg_types.paint ) => attrib( [> `Stroke ]);

let a_stroke_width: wrap( Svg_types.Unit.length ) => attrib( [> `Stroke_Width ]);

let a_stroke_linecap: wrap( [< `Butt | `Round | `Square ] ) => attrib( [> `Stroke_Linecap ]);

let a_stroke_linejoin: wrap( [< `Miter | `Round | `Bever ] ) => attrib( [> `Stroke_Linejoin ]);

let a_stroke_miterlimit: wrap(float) => attrib( [> `Stroke_Miterlimit ]);

let a_stroke_dasharray: wrap(list( Svg_types.Unit.length )) => attrib( [> `Stroke_Dasharray ]);

let a_stroke_dashoffset: wrap( Svg_types.Unit.length ) => attrib( [> `Stroke_Dashoffset ]);

let a_stroke_opacity: wrap(float) => attrib( [> `Stroke_Opacity ]);

Events

Javascript events

let a_onabort: Xml.event_handler => attrib( [> `OnAbort ]);

let a_onactivate: Xml.event_handler => attrib( [> `OnActivate ]);

let a_onbegin: Xml.event_handler => attrib( [> `OnBegin ]);

let a_onend: Xml.event_handler => attrib( [> `OnEnd ]);

let a_onerror: Xml.event_handler => attrib( [> `OnError ]);

let a_onfocusin: Xml.event_handler => attrib( [> `OnFocusIn ]);

let a_onfocusout: Xml.event_handler => attrib( [> `OnFocusOut ]);

let a_onload: Xml.event_handler => attrib( [> `OnLoad ]);

deprecated
Removed in SVG2

let a_onrepeat: Xml.event_handler => attrib( [> `OnRepeat ]);

let a_onresize: Xml.event_handler => attrib( [> `OnResize ]);

let a_onscroll: Xml.event_handler => attrib( [> `OnScroll ]);

let a_onunload: Xml.event_handler => attrib( [> `OnUnload ]);

let a_onzoom: Xml.event_handler => attrib( [> `OnZoom ]);

Javascript mouse events

let a_onclick: Xml.mouse_event_handler => attrib( [> `OnClick ]);

let a_onmousedown: Xml.mouse_event_handler => attrib( [> `OnMouseDown ]);

let a_onmouseup: Xml.mouse_event_handler => attrib( [> `OnMouseUp ]);

let a_onmouseover: Xml.mouse_event_handler => attrib( [> `OnMouseOver ]);

let a_onmouseout: Xml.mouse_event_handler => attrib( [> `OnMouseOut ]);

let a_onmousemove: Xml.mouse_event_handler => attrib( [> `OnMouseMove ]);

let a_ontouchstart: Xml.touch_event_handler => attrib( [> `OnTouchStart ]);

Javascript touch events

let a_ontouchend: Xml.touch_event_handler => attrib( [> `OnTouchEnd ]);

let a_ontouchmove: Xml.touch_event_handler => attrib( [> `OnTouchMove ]);

let a_ontouchcancel: Xml.touch_event_handler => attrib( [> `OnTouchCancel ]);

Elements

let txt: wrap(string) => elt( [> Svg_types.txt ] );

let svg: star ( [< Svg_types.svg_attr ] , [< Svg_types.svg_content ] , [> Svg_types.svg ] ) ;

let g: star ( [< Svg_types.g_attr ] , [< Svg_types.g_content ] , [> Svg_types.g ] ) ;

let defs: star ( [< Svg_types.defs_attr ] , [< Svg_types.defs_content ] , [> Svg_types.defs ] ) ;

let desc: unary ( [< Svg_types.desc_attr ] , [< Svg_types.desc_content ] , [> Svg_types.desc ] ) ;

let title: unary ( [< Svg_types.title_attr ] , [< Svg_types.title_content ] , [> Svg_types.title ] ) ;

let symbol: star ( [< Svg_types.symbol_attr ] , [< Svg_types.symbol_content ] , [> Svg_types.symbol ] ) ;

let use: star ( [< Svg_types.use_attr ] , [< Svg_types.use_content ] , [> Svg_types.use ] ) ;

let image: star ( [< Svg_types.image_attr ] , [< Svg_types.image_content ] , [> Svg_types.image ] ) ;

let switch: star ( [< Svg_types.switch_attr ] , [< Svg_types.switch_content ] , [> Svg_types.switch ] ) ;

let style: unary ( [< Svg_types.style_attr ] , [< Svg_types.style_content ] , [> Svg_types.style ] ) ;

let path: star ( [< Svg_types.path_attr ] , [< Svg_types.path_content ] , [> Svg_types.path ] ) ;

let rect: star ( [< Svg_types.rect_attr ] , [< Svg_types.rect_content ] , [> Svg_types.rect ] ) ;

let circle: star ( [< Svg_types.circle_attr ] , [< Svg_types.circle_content ] , [> Svg_types.circle ] ) ;

let ellipse: star ( [< Svg_types.ellipse_attr ] , [< Svg_types.ellipse_content ] , [> Svg_types.ellipse ] ) ;

let line: star ( [< Svg_types.line_attr ] , [< Svg_types.line_content ] , [> Svg_types.line ] ) ;

let polyline: star ( [< Svg_types.polyline_attr ] , [< Svg_types.polyline_content ] , [> Svg_types.polyline ] ) ;

let polygon: star ( [< Svg_types.polygon_attr ] , [< Svg_types.polygon_content ] , [> Svg_types.polygon ] ) ;

let text: star ( [< Svg_types.text_attr ] , [< Svg_types.text_content ] , [> Svg_types.text ] ) ;

let tspan: star ( [< Svg_types.tspan_attr ] , [< Svg_types.tspan_content ] , [> Svg_types.tspan ] ) ;

let tref: star ( [< Svg_types.tref_attr ] , [< Svg_types.tref_content ] , [> Svg_types.tref ] ) ;

deprecated
Removed in SVG2

let textPath: star ( [< Svg_types.textpath_attr ] , [< Svg_types.textpath_content ] , [> Svg_types.textpath ] ) ;

let altGlyph: unary ( [< Svg_types.altglyph_attr ] , [< Svg_types.altglyph_content ] , [> Svg_types.altglyph ] ) ;

deprecated
Removed in SVG2

type altglyphdef_content = [

| `Ref(list( elt( Svg_types.glyphref )))

| `Item(list( elt( Svg_types.altglyphitem )))

];

let altGlyphDef: unary ( [< Svg_types.altglyphdef_attr ] , [< altglyphdef_content ] , [> Svg_types.altglyphdef ] ) ;

deprecated
Removed in SVG2

let altGlyphItem: star ( [< Svg_types.altglyphitem_attr ] , [< Svg_types.altglyphitem_content ] , [> Svg_types.altglyphitem ] ) ;

deprecated
Removed in SVG2

let glyphRef: nullary ( [< Svg_types.glyphref_attr ] , [> Svg_types.glyphref ] ) ;

deprecated
Removed in SVG2

let marker: star ( [< Svg_types.marker_attr ] , [< Svg_types.marker_content ] , [> Svg_types.marker ] ) ;

let color_profile: star ( [< Svg_types.colorprofile_attr ] , [< Svg_types.colorprofile_content ] , [> Svg_types.colorprofile ] ) ;

deprecated
Removed in SVG2

let linearGradient: star ( [< Svg_types.lineargradient_attr ] , [< Svg_types.lineargradient_content ] , [> Svg_types.lineargradient ] ) ;

let radialGradient: star ( [< Svg_types.radialgradient_attr ] , [< Svg_types.radialgradient_content ] , [> Svg_types.radialgradient ] ) ;

let stop: star ( [< Svg_types.stop_attr ] , [< Svg_types.stop_content ] , [> Svg_types.stop ] ) ;

let pattern: star ( [< Svg_types.pattern_attr ] , [< Svg_types.pattern_content ] , [> Svg_types.pattern ] ) ;

let clipPath: star ( [< Svg_types.clippath_attr ] , [< Svg_types.clippath_content ] , [> Svg_types.clippath ] ) ;

let filter: star ( [< Svg_types.filter_attr ] , [< Svg_types.filter_content ] , [> Svg_types.filter ] ) ;

let feDistantLight: star ( [< Svg_types.fedistantlight_attr ] , [< Svg_types.fedistantlight_content ] , [> Svg_types.fedistantlight ] ) ;

let fePointLight: star ( [< Svg_types.fepointlight_attr ] , [< Svg_types.fepointlight_content ] , [> Svg_types.fepointlight ] ) ;

let feSpotLight: star ( [< Svg_types.fespotlight_attr ] , [< Svg_types.fespotlight_content ] , [> Svg_types.fespotlight ] ) ;

let feBlend: star ( [< Svg_types.feblend_attr ] , [< Svg_types.feblend_content ] , [> Svg_types.feblend ] ) ;

let feColorMatrix: star ( [< Svg_types.fecolormatrix_attr ] , [< Svg_types.fecolormatrix_content ] , [> Svg_types.fecolormatrix ] ) ;

let feComponentTransfer: star ( [< Svg_types.fecomponenttransfer_attr ] , [< Svg_types.fecomponenttransfer_content ] , [> Svg_types.fecomponenttransfer ] ) ;

let feFuncA: star ( [< Svg_types.fefunca_attr ] , [< Svg_types.fefunca_content ] , [> Svg_types.fefunca ] ) ;

let feFuncG: star ( [< Svg_types.fefuncg_attr ] , [< Svg_types.fefuncg_content ] , [> Svg_types.fefuncg ] ) ;

let feFuncB: star ( [< Svg_types.fefuncb_attr ] , [< Svg_types.fefuncb_content ] , [> Svg_types.fefuncb ] ) ;

let feFuncR: star ( [< Svg_types.fefuncr_attr ] , [< Svg_types.fefuncr_content ] , [> Svg_types.fefuncr ] ) ;

let feComposite: star ( [< Svg_types.fecomposite_attr ] , [< Svg_types.fecomposite_content ] , [> Svg_types.fecomposite ] ) ;

let feConvolveMatrix: star ( [< Svg_types.feconvolvematrix_attr ] , [< Svg_types.feconvolvematrix_content ] , [> Svg_types.feconvolvematrix ] ) ;

let feDiffuseLighting: star ( [< Svg_types.fediffuselighting_attr ] , [< Svg_types.fediffuselighting_content ] , [> Svg_types.fediffuselighting ] ) ;

let feDisplacementMap: star ( [< Svg_types.fedisplacementmap_attr ] , [< Svg_types.fedisplacementmap_content ] , [> Svg_types.fedisplacementmap ] ) ;

let feFlood: star ( [< Svg_types.feflood_attr ] , [< Svg_types.feflood_content ] , [> Svg_types.feflood ] ) ;

let feGaussianBlur: star ( [< Svg_types.fegaussianblur_attr ] , [< Svg_types.fegaussianblur_content ] , [> Svg_types.fegaussianblur ] ) ;

let feImage: star ( [< Svg_types.feimage_attr ] , [< Svg_types.feimage_content ] , [> Svg_types.feimage ] ) ;

let feMerge: star ( [< Svg_types.femerge_attr ] , [< Svg_types.femerge_content ] , [> Svg_types.femerge ] ) ;

let feMorphology: star ( [< Svg_types.femorphology_attr ] , [< Svg_types.femorphology_content ] , [> Svg_types.femorphology ] ) ;

let feOffset: star ( [< Svg_types.feoffset_attr ] , [< Svg_types.feoffset_content ] , [> Svg_types.feoffset ] ) ;

let feSpecularLighting: star ( [< Svg_types.fespecularlighting_attr ] , [< Svg_types.fespecularlighting_content ] , [> Svg_types.fespecularlighting ] ) ;

let feTile: star ( [< Svg_types.fetile_attr ] , [< Svg_types.fetile_content ] , [> Svg_types.fetile ] ) ;

let feTurbulence: star ( [< Svg_types.feturbulence_attr ] , [< Svg_types.feturbulence_content ] , [> Svg_types.feturbulence ] ) ;

let cursor: star ( [< Svg_types.cursor_attr ] , [< Svg_types.cursor_content ] , [> Svg_types.cursor ] ) ;

let a: star ( [< Svg_types.a_attr ] , [< Svg_types.a_content ] , [> Svg_types.a ] ) ;

let view: star ( [< Svg_types.view_attr ] , [< Svg_types.view_content ] , [> Svg_types.view ] ) ;

let script: unary ( [< Svg_types.script_attr ] , [< Svg_types.script_content ] , [> Svg_types.script ] ) ;

let animation: star ( [< Svg_types.animation_attr ] , [< Svg_types.animation_content ] , [> Svg_types.animation ] ) ;

let set: star ( [< Svg_types.set_attr ] , [< Svg_types.set_content ] , [> Svg_types.set ] ) ;

let animateMotion: star ( [< Svg_types.animatemotion_attr ] , [< Svg_types.animatemotion_content ] , [> Svg_types.animatemotion ] ) ;

let mpath: star ( [< Svg_types.mpath_attr ] , [< Svg_types.mpath_content ] , [> Svg_types.mpath ] ) ;

let animateColor: star ( [< Svg_types.animatecolor_attr ] , [< Svg_types.animatecolor_content ] , [> Svg_types.animatecolor ] ) ;

let animateTransform: star ( [< Svg_types.animatetransform_attr ] , [< Svg_types.animatetransform_content ] , [> Svg_types.animatetransform ] ) ;

let font: star ( [< Svg_types.font_attr ] , [< Svg_types.font_content ] , [> Svg_types.font ] ) ;

deprecated
Removed in SVG2

let glyph: star ( [< Svg_types.glyph_attr ] , [< Svg_types.glyph_content ] , [> Svg_types.glyph ] ) ;

deprecated
Removed in SVG2

let missing_glyph: star ( [< Svg_types.missingglyph_attr ] , [< Svg_types.missingglyph_content ] , [> Svg_types.missingglyph ] ) ;

deprecated
Removed in SVG2

let hkern: nullary ( [< Svg_types.hkern_attr ] , [> Svg_types.hkern ] ) ;

deprecated
Removed in SVG2

let vkern: nullary ( [< Svg_types.vkern_attr ] , [> Svg_types.vkern ] ) ;

deprecated
Removed in SVG2

let font_face: nullary ( [< Svg_types.font_face_attr ] , [> Svg_types.font_face ] ) ;

deprecated
Removed in SVG2

let font_face_src: star ( [< Svg_types.font_face_src_attr ] , [< Svg_types.font_face_src_content ] , [> Svg_types.font_face_src ] ) ;

deprecated
Removed in SVG2

let font_face_uri: star ( [< Svg_types.font_face_uri_attr ] , [< Svg_types.font_face_uri_content ] , [> Svg_types.font_face_uri ] ) ;

deprecated
Removed in SVG2

let font_face_format: nullary ( [< Svg_types.font_face_format_attr ] , [> Svg_types.font_face_format ] ) ;

deprecated
Removed in SVG2

let font_face_name: nullary ( [< Svg_types.font_face_name_attr ] , [> Svg_types.font_face_name ] ) ;

deprecated
Removed in SVG2

let metadata: ?a:list( attrib( Svg_types.metadata_attr )) => list_wrap( Xml.elt ) => elt( [> Svg_types.metadata ] );

let foreignObject: ?a:list( attrib( Svg_types.foreignobject_attr )) => list_wrap( Xml.elt ) => elt( [> Svg_types.foreignobject ] );

Deprecated

let pcdata: wrap(string) => elt( [> Svg_types.txt ] );

deprecated
Use txt instead

Conversion with untyped representation

WARNING: These functions do not ensure HTML or SVG validity! You should always explicitly given an appropriate type to the output.

let of_seq: Stdlib.Seq.t( Xml_stream.signal ) => list_wrap( elt( 'a));

import signal converts the given XML signal into Tyxml elements. It can be used with HTML and SVG parsing libraries, such as Markup.

raises Xml_stream.Malformed_stream
if the stream is malformed.

let tot: Xml.elt => elt( 'a);

let totl: list_wrap( Xml.elt ) => list_wrap( elt( 'a));

let toelt: elt( 'a) => Xml.elt ;

let toeltl: list_wrap( elt( 'a)) => list_wrap( Xml.elt );

let doc_toelt: doc => Xml.elt ;

let to_xmlattribs: list(attrib ('a)) => list( Xml.attrib );

let to_attrib: Xml.attrib => attrib( 'a);

module Unsafe : { ... };

Unsafe features.