Module Tyxml_name

Gives the tyxml names for HTML elements and attributes.

let to_ocaml: string => string;

The transformations are the following:

Valid letters in OCaml identifiers are kept.

Everything else is turn into an underscore '_'.

let ident: string => string;

Turn the given element name into a valid identifier.

Follow the to_ocaml convention and lowercase the first letter.

let attrib: string => string;

Turn the given attribute name into a valid identifier.

Follow the to_ocaml convention and add "a_" at the beginning.

let polyvar: string => string;

Turn the given name into a valid Polymorphic variant name.

Follow the to_ocaml convention, uppercase the first letter and add "`".