Module Tyxml_svg.Xml

Underlying XML data-structure

The type variables in elt and attrib are know as phantom types. The implementation, defined here, is actually monomorphic.

In particular, tyxml doesn't impose any overhead over the underlying representation. The tot and toelt functions allows to convert between the typed and the untyped representation without any cost.

Note that some implementation may not be iterable or printable, such as the Dom representation exposed by js_of_ocaml.

module W = Xml_wrap.NoWrap ;

type wrap('a) = W.t ('a) ;

type list_wrap('a) = W.tlist ('a) ;

type uri = Tyxml_xml.uri ;

let string_of_uri: W.ft (uri, string) ;

let uri_of_string: W.ft (string, uri) ;

type aname = string;

type event_handler = Tyxml_xml.event_handler ;

type mouse_event_handler = Tyxml_xml.mouse_event_handler ;

type keyboard_event_handler = Tyxml_xml.keyboard_event_handler ;

type touch_event_handler = Tyxml_xml.touch_event_handler ;

type attrib = Tyxml_xml.attrib ;

let float_attrib: aname => wrap(float ) => attrib ;

let int_attrib: aname => wrap(int) => attrib ;

let string_attrib: aname => wrap( string) => attrib ;

let space_sep_attrib: aname => wrap(list( string)) => attrib ;

let comma_sep_attrib: aname => wrap(list( string)) => attrib ;

let event_handler_attrib: aname => event_handler => attrib ;

let mouse_event_handler_attrib: aname => mouse_event_handler => attrib ;

let keyboard_event_handler_attrib: aname => keyboard_event_handler => attrib ;

let touch_event_handler_attrib: aname => touch_event_handler => attrib ;

let uri_attrib: aname => wrap( uri) => attrib ;

let uris_attrib: aname => wrap(list( uri)) => attrib ;

type elt = Tyxml_xml.elt ;

type ename = string;

let empty: unit => elt;

let comment: string => elt;

let pcdata: wrap( string) => elt;

let encodedpcdata: wrap( string) => elt;

let entity: string => elt;

let leaf: ?a:list( attrib) => ename => elt;

let node: ?a:list( attrib) => ename => list_wrap( elt) => elt;

let cdata: string => elt;

let cdata_script: string => elt;

let cdata_style: string => elt;