Module Tyxml_xml

Basic functions for construction and manipulation of XML tree.

include Xml_sigs.Iterable with type uri = string and type event_handler = string and type mouse_event_handler = string and type keyboard_event_handler = string and type touch_event_handler = string ;

include Xml_sigs.NoWrap with type uri = string with type event_handler = string with type mouse_event_handler = string with type keyboard_event_handler = string with type touch_event_handler = string ;

module W = Xml_wrap.NoWrap ;

type wrap('a) = W.t ('a) ;

type list_wrap('a) = W.tlist ('a) ;

type uri = string;

let string_of_uri: W.ft (uri, string) ;

let uri_of_string: W.ft (string, uri) ;

type aname = string;

type event_handler = string;

type mouse_event_handler = string;

type keyboard_event_handler = string;

type touch_event_handler = string;

type attrib ;

let float_attrib: aname => wrap(float ) => attrib ;

let int_attrib: aname => wrap(int) => attrib ;

let string_attrib: aname => wrap( string) => attrib ;

let space_sep_attrib: aname => wrap(list( string)) => attrib ;

let comma_sep_attrib: aname => wrap(list( string)) => attrib ;

let event_handler_attrib: aname => event_handler => attrib ;

let mouse_event_handler_attrib: aname => mouse_event_handler => attrib ;

let keyboard_event_handler_attrib: aname => keyboard_event_handler => attrib ;

let touch_event_handler_attrib: aname => touch_event_handler => attrib ;

let uri_attrib: aname => wrap( uri) => attrib ;

let uris_attrib: aname => wrap(list( uri)) => attrib ;

type elt ;

type ename = string;

let empty: unit => elt;

let comment: string => elt;

let pcdata: wrap( string) => elt;

let encodedpcdata: wrap( string) => elt;

let entity: string => elt;

let leaf: ?a:list( attrib) => ename => elt;

let node: ?a:list( attrib) => ename => list_wrap( elt) => elt;

let cdata: string => elt;

let cdata_script: string => elt;

let cdata_style: string => elt;

type separator =

| Space

| Comma

;

let aname: attrib => aname;

type acontent = pri

| AFloat(float)

| AInt(int)

| AStr(string)

| AStrL( separator , list(string))

;

let acontent: attrib => acontent;

type econtent = pri

| Empty

| Comment(string)

| EncodedPCDATA(string)

| PCDATA(string)

| Entity(string)

| Leaf( ename, list( attrib))

| Node( ename, list( attrib), list (elt))

;

let content: elt => econtent;

include Xml_sigs.Pp with type elt := elt ;

let pp: ?encode: (string => string) => ?indent:bool => unit => Stdlib.Format.formatter => elt => unit;

pp () is a Format printer for untyped XML.

It can be used in combination with "%a". For example, to get a string:

        let s = Format.asprintf "%a" (pp ()) my_xml
       

A custom encoding function can be provided with the ~encode argument. Various implementations of encode are available in Xml_print .

Import/Export

let of_seq: Stdlib.Seq.t( Xml_stream.signal ) => list(elt);

Iterators

let amap: ( ename => list( attrib) => list( attrib)) => elt => elt;

Recursively edit attributes for the element and all its children.

let amap1: ( ename => list( attrib) => list( attrib)) => elt => elt;

Edit attributes only for one element.

The following can safely be exported by higher level libraries, because removing an attribute from a element is always legal.

let rm_attrib: ( aname => bool) => list(attrib) => list( attrib);

let rm_attrib_from_list: ( aname => bool) => (string => bool) => list(attrib) => list( attrib);

let map_int_attrib: ( aname => bool) => (int => int) => list(attrib) => list( attrib);

let map_string_attrib: ( aname => bool) => (string => string) => list(attrib) => list( attrib);

let map_string_attrib_in_list: ( aname => bool) => (string => string) => list(attrib) => list( attrib);

Exporting the following by higher level libraries would drive a hole through a type system, because they allow to add any attribute to any element.

let add_int_attrib: aname => int => list(attrib) => list( attrib);

let add_string_attrib: aname => string => list(attrib) => list( attrib);

let add_comma_sep_attrib: aname => string => list(attrib) => list( attrib);

let add_space_sep_attrib: aname => string => list(attrib) => list( attrib);

let fold: (unit => 'a) => (string => 'a) => (string => 'a) => (string => 'a) => (string => 'a) => ( ename => list( attrib) => 'a) => ( ename => list( attrib) => list('a) => 'a) => elt => 'a;

let all_entities: elt => list(string);

let translate: ( ename => list( attrib) => elt) => ( ename => list( attrib) => list(elt) => elt) => ( 'state => ename => list( attrib) => list(elt)) => ( 'state => ename => list( attrib) => list(elt) => list(elt)) => ( ename => list( attrib) => 'state => 'state) => 'state => elt => elt;

Deprecated printers

let print_list: output: (string => unit) => ?encode: (string => string) => list(elt) => unit;

let print: Stdlib.Format.formatter => elt => unit;