Module Xml_print.Make_typed_fmt

Printers for typed XML modules such as the one produced by Svg_f and Html_f .

Parameters

module Xml : Xml_sigs.Iterable

module Typed_xml : Xml_sigs.Typed_xml with module Xml := Xml

Signature

let pp_elt: ?encode: (string => string) => ?indent:bool => unit => Stdlib.Format.formatter => Typed_xml.elt ('a) => unit;

pp_elt () is a Format printer for individual elements.

A custom encoding function can be provided with the ~encode argument. Various implementations of encode are available in Xml_print .

let pp: ?encode: (string => string) => ?indent:bool => ?advert:string => unit => Stdlib.Format.formatter => Typed_xml.doc => unit;

pp () is a Format printer for complete documents.

It can be used in combination with "%a". For example, to get a string:

      let s = Format.asprintf "%a" (pp ()) my_document
     

A custom encoding function can be provided with the ~encode argument. Various implementations of encode are available in Xml_print .